フォームの要素をコンポーネントに分割
