+++ src/components/index.ts
@@ -4,7 +4,12 @@ export { SearchInputGrey } from './atoms/SearchInputGrey';
 export { SearchInput } from './molecules/SearchInput';
 export { SelectOrder } from './molecules/SelectOrder';
 export { SignupDialog } from './molecules/SignupDialog';
+export { SignupFormName } from './atoms/SignupFormName';
+export { SignupFormEmail } from './atoms/SignupFormEmail';
+export { SignupFormPassword } from './atoms/SignupFormPassword';
 export { LoginDialog } from './molecules/LoginDialog';
+export { LoginFormEmail } from './atoms/LoginFormEmail';
+export { LoginFormPassword } from './atoms/LoginFormPassword';
 export { StyledDateField } from './custom/StyledDateField';
 export { StyledInput } from './custom/StyledInput';
 export { StyledSelectMenuItem as StyledMenuItem } from './custom/StyledSelectMenuItem';
